Created enclosure for ideal diode
The ideal diode that should be between the house battery and start battery. Since only the house battery is charged with solar panels, the solar panel will now also charge the start battery, though it will always make sure the the start battery is more charged than the house battery.